rabbitMQ_api/requirements.txt
pika
Flask==1.1.1
Flask-Threads
python-dotenv
pandas
mysql
numpy